+++ src/storage/store_lock/store_lock_async/default_async.rs
@@ -30,7 +30,7 @@ impl AsyncStoreKeyMutexTraits for AsyncDefaultStoreMutex {
 }
 
 /// Default store locks.
-#[derive(Debug, Default, derive_more::Constructor)]
+#[derive(Debug, Default)]
 pub struct AsyncDefaultStoreLocks(Mutex<HashMap<StoreKey, Arc<Mutex<()>>>>);
 
 #[async_trait::async_trait]
@@ -40,5 +40,39 @@ impl AsyncStoreLocksTraits for AsyncDefaultStoreLocks {
         Box::new(AsyncDefaultStoreMutex(
             locks.entry(key.clone()).or_default().clone(),
         ))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{sync::atomic::AtomicUsize, time::Duration};
+
+    use crate::storage::{store::AsyncObjectStore, AsyncReadableWritableStorageTraits};
+
+    use super::*;
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+    async fn store_default_lock_async() {
+        let store = Arc::new(AsyncObjectStore::new_with_locks(
+            object_store::memory::InMemory::default(),
+            Arc::new(AsyncDefaultStoreLocks::default()),
+        ));
+        let locks_held = Arc::new(AtomicUsize::new(0));
+        let futures = (0..20).into_iter().map(|_| {
+            let key = StoreKey::new("key").unwrap();
+            let store = store.clone();
+            let locks_held = locks_held.clone();
+            tokio::task::spawn(async move {
+                let mutex = store.mutex(&key).await.unwrap();
+                let _lock = mutex.lock().await;
+                locks_held.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
+                std::thread::sleep(Duration::from_millis(10));
+                let locks_held = locks_held.fetch_sub(1, std::sync::atomic::Ordering::SeqCst);
+                locks_held == 1
+            })
+        });
+        let result = futures::future::try_join_all(futures).await.unwrap();
+        println!("{result:?}");
+        assert!(result.iter().all(|b| *b));
     }
 }