+++ src/storage/store_lock/store_lock_sync/default_sync.rs
@@ -29,7 +29,7 @@ impl StoreKeyMutexTraits for DefaultStoreMutex {
 }
 
 /// Default store locks.
-#[derive(Debug, Default, derive_more::Constructor)]
+#[derive(Debug, Default)]
 pub struct DefaultStoreLocks(Mutex<HashMap<StoreKey, Arc<Mutex<()>>>>);
 
 impl StoreLocksTraits for DefaultStoreLocks {
@@ -38,5 +38,31 @@ impl StoreLocksTraits for DefaultStoreLocks {
         Box::new(DefaultStoreMutex(
             locks.entry(key.clone()).or_default().clone(),
         ))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{sync::atomic::AtomicUsize, time::Duration};
+
+    use rayon::iter::{IntoParallelIterator, ParallelIterator};
+
+    use crate::storage::{store::MemoryStore, ReadableWritableStorageTraits};
+
+    use super::*;
+
+    #[test]
+    fn store_default_lock_sync() {
+        let store = MemoryStore::new_with_locks(Arc::new(DefaultStoreLocks::default()));
+        let key = StoreKey::new("key").unwrap();
+        let locks_held = AtomicUsize::new(0);
+        (0..20).into_par_iter().for_each(|_| {
+            let mutex = store.mutex(&key).unwrap();
+            let _lock = mutex.lock();
+            locks_held.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
+            std::thread::sleep(Duration::from_millis(10));
+            let locks_held = locks_held.fetch_sub(1, std::sync::atomic::Ordering::SeqCst);
+            assert_eq!(locks_held, 1);
+        });
     }
 }